refactor: Relocate TitledContainer and remove unused date utility functions (#203)
- Moved `TitledContainer` from `common.component` to `presentation.rentaldetail.components`
- Removed unused `DEPOSIT_BASIS_DAYS` constant
- Removed redundant `toShortFormat()` extension for `LocalDateTime`
- Removed unused string-based overload of `inclusiveDaysBetween`
