+++ app/src/main/java/com/example/rentit/common/Constants.kt
@@ -1,5 +1,4 @@
 package com.example.rentit.common
 
-const val DEPOSIT_BASIS_DAYS = 3
 const val D_DAY_ALERT_THRESHOLD_DAYS = 3
 const val PRICE_LIMIT = 5_000_000
