+++ app/src/main/java/com/example/rentit/common/util/DateFormat.kt
@@ -3,13 +3,8 @@ package com.example.rentit.common.util
 import android.os.Build
 import androidx.annotation.RequiresApi
 import java.time.LocalDate
-import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
 @RequiresApi(Build.VERSION_CODES.O)
 fun LocalDate.toShortFormat(): String =
-    this.format(DateTimeFormatter.ofPattern("yy.MM.dd"))
-
-@RequiresApi(Build.VERSION_CODES.O)
-fun LocalDateTime.toShortFormat(): String =
     this.format(DateTimeFormatter.ofPattern("yy.MM.dd"))
