+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/components/TitledContainer.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.common.component
+package com.example.rentit.presentation.rentaldetail.components
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -17,6 +17,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.rentit.common.component.screenHorizontalPadding
 import com.example.rentit.common.theme.AppBlack
 import com.example.rentit.common.theme.Gray100
 import com.example.rentit.common.theme.RentItTheme
