+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/components/section/RentalTrackingSection.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.rentit.R
 import com.example.rentit.common.theme.RentItTheme
-import com.example.rentit.common.component.TitledContainer
+import com.example.rentit.presentation.rentaldetail.components.TitledContainer
 import com.example.rentit.common.theme.AppBlack
 import com.example.rentit.common.theme.Gray400
 import com.example.rentit.presentation.rentaldetail.components.LabeledValue
